+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 # Changelog
 
 ## V8.0
-- Initial Music Player Module.
+- Initial Playlists Module.